add pick up and fixes

+++ src/conversions.cpp
@@ -1,7 +1,7 @@
 #include "conversions.h"
 
 b2Vec2 convert(const Tmx::Point& from, Map* const map){
-	return b2Vec2(from.x / map->getTileWidth(), from.y / map->getTileHeight());
+	return b2Vec2((float)from.x / map->getTileWidth(), (float)from.y / map->getTileHeight());
 }
 
 b2Vec2 convert(const sf::Vector2f& from, shared_ptr<Map> map){
